+++ src/utils/random.rs
@@ -3,9 +3,7 @@ use std::cell::RefCell;
 
 use ahash::RandomState;
 
-use crate::models::FloatType;
-
-#[derive(Debug)]
+#[derive(Clone, Debug)]
 pub struct Random {
     pub rng: RefCell<Rng>,
     pub seed: Option<u64>,
@@ -30,8 +28,8 @@ impl Random {
         self.rng.borrow_mut().shuffle(container);
     }
 
-    pub fn real(&self) -> FloatType {
-        self.rng.borrow_mut().f64() as FloatType
+    pub fn real(&self) -> f64 {
+        self.rng.borrow_mut().f64() as f64
     }
 
     // Get random number in range [lower, upper). Upper is not inclusive
